Fix dangling pointer in xwalk extension code.
At the time we reach OnRenderProcessHostClosed the IPC messaging filters
are already deleted because the IPC Channel has been closed. We do not
need to invalidate the ExtensionServerMessageFilter instance because
it's already deleted (and the various overloads of MessageFilter are
going to invalidate it) and we should not remove from the current
channel because it's been already deleted (removing it will cause its
deletion). In fact we do not need to hold a reference of that message
filter as it's own by the IPC channel and will be deleted.

BUG=XWALK-4295