Backport patch to fix an old skia regression
Backport https://codereview.chromium.org/1161063003 which fixes
an old Skia performance regression on popular devices like Samsung
Galaxy S3. The above CL was landed in M45 and therefore will be
part in our next rebase (it was first landed, then reverted, then landed
again after some additional fixes). However this regression is a pressing i
ssue for some stakeholders so they can move to newer versions of Crosswalk
(as opposed to stick with crosswalk 7 the last version where there
is no performance regression). As we do not fork Skia repository
the proposed solution is to apply a patch when running gclient. That
way it is easier to remove when rebasing to M45. Please note that the
provided patch file has been modified to apply on M44.

BUG=XWALK-3720