Add code to build Debian desktop packages.
This work is still experimental, so here be dragons!

The whole patch is inspired by the code in Chromium itself in
`src/chrome/installer/linux` that is used to generate RPMs and Debian
packages.

Consequently, this means that at least for now we are not following
Debian's best practices:
- We do not build from a source tarball; instead, the new gyp target
  `xwalk_installer_linux_debian` creates a package out of the files
  generated by the `xwalk` target.
- To achieve that, we make some fairly low-level calls to `dpkg-deb`
  directly.

We may revisit the way we create packages in the future; for now, this
is the quickest way to start having Debian packages and also saves us
some work if/when we decide to ship NaCl's support files, as we do not
need to care about downloading all the toolchain tarball files ourselves
since it is done by `gclient runhooks'.

TODO items:
- Have a `copyright` file.
- Consider whether we should install into `/usr` directly.
- Possibly switch to a more standard Debian packaging workflow.
- Ship NaCl's support files.

BUG=XWALK-4374