Work around Skia performance regression
A Skia CL (https://codereview.chromium.org/607993002 ) made CrossWalk
regress on some devices such as Galaxy S3. Upstream proposed a CL to fix
the performance regression in
https://codereview.chromium.org/1161063003. Alexis backported upstream's
fix to CrossWalk master in
https://github.com/crosswalk-project/crosswalk/pull/3096. But it's hard
to backport to CrossWalk 14 due to many dependency issues in Skia
repo. Not only the real fix is needed to backport but also other
dependent CLs are needed to backport. So I propose to revert the origin
Skia CL which introduced the regression as a workaroud for CrossWalk 14.

Before the regressed CL, Skia use a fallback draw as the first choice
for surface copy.  In this CL I remove other copy methods before the
fallback draw. This is the same logic as before. Also upstream's fix did
the same thing: https://codereview.chromium.org/1161063003. With this
CL, performance of CrossWalk is back. For a test case (propaganda: a
game from community), it can achieve 55FPS compared to previous
5FPS.