[Android] Deprecate XWalkApplication
XWalkApplication's role is to help the app to get resources from the
XWalkRuntimeLib.apk in shared mode, but it also cause the
inconvenience of developers. So we integrate the whole resources to
the app even in shared mode to get rid of this class.
(cherry picked from commit 7748996639c1e1292ffafd43dc6c60ad01d56c49)