Presentation API Reimplementation
Modified the implementation of Presentation API on Android, according
to the new version of W3C Presentation API Spec:
http://www.w3.org/TR/presentation-api/

The major change of the new implementation is to introduced the
session object, which is responsible for representing a presentation
session, and setup a message channel between host screen and remote
display screen. And the major API (e.g. startSession) returns Promise
object instead of callback functions.

BUG=XWALK-4646