[Windows][Extensions]Fix render process crash when sending data from native to JS using the shared memory handle.
This is the second fix for this bug. The previous fix uncovered this
one.

The crash was because the memory handle we send to the extension process
was null. The fact it was null is because on Windows it was failing to
share the handle of the shared memory between the render process and
the extension process. GiveReadOnlyToProcess was failing inside
windows specific code, more exactly in DuplicateHandle, a win API
which was returning an error code of 6. 6 means the handle we're passing
is invalid. Turns out that while we store the renderer process handle
earlier (so we can share the data between the two), the open() call
on the process is too early. It seems like Windows can't work with this
handle unless the process is open right before. Moving the open() call
to the last moment (by storing the PID instead of the handle) doesn't
fully work (we don't fail with error 6 but with 5 which means there
is not enough privileges to open/share with that process). We therefore
open it with extra privileges which means we can share data between
the two. I also added a LOG and an early return when sharing the handle
doesn't work, after all it means it's functionally broken.

While investigating this issue it turns out that the automatic test was
broken because it was always reporting "Pass" even if the actual message
passing with data handle was broken. The fix consists of moving it from
a regular extension test to a true external extension test. The second
part is also making sure we don't report "Pass" until we received all
the messages back from the native extension. If we don't receive one
the test will time out. I've decided to address it in this patch so that
I don't need to fix the test, then disable it to land that patch and
re-enable it again.

It turns out that the fix made the test to fail on Linux (running the test
into a regular xwalk instance was working fine) because the process
couldn't be opened. After debbuging it turns out that the test was
run in a in mode where the extensions are not created the same way as
an external extension (created as in process). This will make our code
for sharing memory to fail because there is no multi-process involved.
The test never worked but the code worked, it was discovered together
with this patch.

BUG=XWALK-4295