Preserve local storage when migrating from webview to Crosswalk
Copy the data of web storage directory to Crosswalk profile name for
Android Kitkat.
The KK webview and the pre-KK webview use different formate to store data,
but the storage database have implemented the mechanism that it will be upgraded
to V2(BLOB value column) if there was V1 (TEXT value column) database already.

BUG=XWALK-4349