[Android] Fix the crash issue when running file transfer.
This patch is to fix the crash issue when running file transfer tests in
mobilespeck.
The exposed APIs were not initialized when the explicit constructor was
not existed in the internal java file.
Add the default constructor to call init function.

BUG=XWALK-4667

(cherry picked from commit 2c0aaff4a3807c2e012bca572789b5338d690867)