[Android] Rewrite merge_jars.py.
Almost all the code that was present in the script is also available in
Chromium itself, so it makes no sense to have the uptenth
reimplementation of GetCommandOutput() or calls to jar(1).

Additionally, the handling of the --jars option was broken: it always
assumed that each path would be properly escaped by gyp and could be
eval()ed directly into Python. Not only does calling eval() in this
context make zero sense, but it also fails with targets such as
`android_support_v13_javalib`, which is passed as an absolute path
without escaped shell quotes. This is a problem starting with M45, as
that target is depended upon by `content_java`.

The --classes-dir option is also useless: instead of creating and
cleaning directories inside the build directory, just use a different
temporary directory each time.

The only difference in the generated jar files is the presence of a
temporary directory with an `.empty` file there that comes from the code
in src/build/android/gyp/jar.py. It should not cause any trouble.