Remove use of MessageLoopProxy and deprecated MessageLoop APIs.
Chromium is deprecating MessageLoopProxy and moving towards the
TaskRunner API instead. Bug 465354 contains more information.

Since we can already do this in M44, do it separately from moving to M45
so that the size of the M45 patch is smaller. Thanks to Olli Raula and
Mrunal Kapade for doing the majority of the work.

Most of the changes are mechanic:
* base::MessageLoopProxy::current() becomes
  base::ThreadTaskRunnerHandle::Get().
* scoped_refptr<base::MessageLoopProxy> becomes
  scoped_refptr<base::SingleThreadTaskRunner>.
* base::Thread::message_loop_proxy() becomes
  base::Thread::task_runner().

The only noteworthy change is in XWalkExtensionFunctionHandler, caused
by the fact that ThreadTaskRunnerHandle::Get() enforces an assertion
that was commented out in MessageLoopProxy. This assertion makes sure
the call is being performed in a thread with a valid task runner.

That is not the case for XWalkExtensionFunctionHandlerTest's
PostingResultAfterDeletingTheHandler. Its call to XWEFH::HandleMessage()
is done without a runner, so we need to check before calling
ThreadTaskRunnerHandle::Get(). This is similar to Chromium commits
afd978c6 ("gpu: Fix webview parent compositor crash in context loss")
and 0ad2c54d ("gpu: Remove use of MessageLoopProxy and deprecated
MessageLoop APIs").