[Tizen] Remove support for pausing and resuming the JS engine.
This (manually) reverts the following commits:
- 94d0859a0920b991d4294e8ab8969e61bc734888
  [Tizen][Runtime] Enable pause/resume web application JS engine
- 3bf14f044e46f9ebe772b004a34c33eb75c201f9
  [Tizen] Enable background-support attribute in settings element
- 01be04ccda9da53680a8e723278aa6eb5d068e41
  [Tizen] Use appcore RESET event as application starting point

All the code relied upon suspending Blink's shared timer, which has been
removed in M45 -- see upstream bug 463143 for all related commits and
rationale.

Prepare for M45 by removing the code separately so that the change is
more visible and easier to revert in the future should we ever implement
support for it differently.

BUG=XWALK-1148
BUG=XWALK-1497
BUG=XWALK-1498
BUG=XWALK-2779