[Android] Add support disabling copy/paste in XWalkView
Delegate ContentView's performLongClick() to XWalkView so that developer
can override XWalkView's performLongClick() or setOnLongClickListener() to
disable copy/paste action bar.
A new annotation XWalkAPI.delegate is used to ensure that the callback
from XWalkViewInternal to XWalkView is one-way and the actual implementation
of callback in XWalkView side is specified via XWalkAPI.preWrapperLines.

BUG=XWALK-4786