Implement .NET external extension system.
This commit adds support to load external .NET extensions
for 3rd party developers.

You can specify the path of the extensions to load using
--external-extension-path=/path/to/extension

The extensions are expected to be a .DLL constructed with
Visual Studio as a C# Library (and thus multiarch). Alongside
with that library the 3rd party developer needs to include a .NET bridge
named myextension_bridge.dll given the .NET extension is called
myextension.dll. The dotnet bridge is shipped part of Crosswalk
and is called xwalk_dotnet_bridge and it's a simple copy/rename
for the developer.

This .NET library needs to contains two classes XWalkExtension and
XWalkExtensionInstance which are going to be invoked by Crosswalk
to load the extensions and call the various methods on each instances
to make communication between .NET and JS. .NET extensions are a bit
higher level than our C or C++ extensions because the .NET extension
system will automatically call the constructors in .NET when needed :
e.g. creating the XWalkExtension instance and then create the various
XWalkExtensionInstance instances. This is following inspiration of
what's inside tizen-extension-crosswalk common/ directory which is used
to abstract the low levelness to who write t-e-c extensions.

 _______       _______       ______________
| xwalk | <-> | bridge| <-> |.NET extension|
|_______|     |_______|     |______________|

The bridge is loaded dynamically like an external extension would
be loaded : it's a DLL and it contains a single exported symbol : XW_Initialize.

One bridge instance is used by each extension and it is responsible to load
the .NET extension, verify using .NET reflection that the entry points
(e.g. XWalkExtension constructor, XWalkInstance handleMessage method...)
are present and that the .NET extension can be used. It then handles the
communication back and forth with Crosswalk and the extension.

One implementation detail worth to note is the usage of the dynamic parameter
in XWalkExtensionInstance, this pointer is used to communicate back from
.NET to C++/XWalk and the fact is dynamic is to hide implementation
details to the 3rd party developer which should only care about calling
the right functions on it. If they don't exists, there will be a runtime
error showing up. It makes using the system very simple (nothing to
include, nothing to import).

Please note that we don't support all the features of the extension
system at the moment, for example you can't add multiple entry points
in JS, and you can't have access to the runtime variables inside the
extension. These features can be added in a following patch.

This commit also roll GYP forward so we can compile passing /clr as
a compiler paramenter. It will probably be not needed by the next
rebase.

Part of the feature work I've added a bunch of tests which
actually build .NET extensions and test the overall system (thus
the .csproj files checked in). There is a little script to invoke
msbuild so it builds the DLL in .NET (GYP doesn't know about that).