Fix build on Windows.
And most likely on Mac as well.

https://github.com/crosswalk-project/crosswalk/pull/320 broke the
build by using classes not added to the build.

Fix the build by adding the new class from #320 to the build but it's
most likely that the Download feature doesn't yet work on Windows
because there are still code guarded in #ifdef in the original
commit. I've created XWALK-4969 to enable it which shouldn't require
much work.