Roll Chromium 46.0.2490.86
This is much delayed stable M46 release so thanks for your patience.
Big thanks to Alexis Menard, Jon Willis, Lin Sun and Chuan Liu for all
their contributions towards making this release buildable and functional.
Last but not least, thanks to Raphael kubo Da Costa for all the support
provided for this release.

Changes and notes:
* SIMD.js: This release starts using the upstream implementation of SIMD.js.
You will notice the js flag being changed from --simd-object to --harmony-simd.
Since this is the first release after SIMD was upstreamed some of the performance
optimization which were part of earlier releases are not available in this release.

Upstream CLs we had to adapt to (not a complete list):

https://codereview.chromium.org/1242023005
https://codereview.chromium.org/1249133002
https://codereview.chromium.org/1281323003
https://codereview.chromium.org/1242953004
https://codereview.chromium.org/1295963005
https://codereview.chromium.org/1280483004
https://codereview.chromium.org/1164073005
https://codereview.chromium.org/1277213003
https://codereview.chromium.org/1272113002
https://codereview.chromium.org/1242673002
https://codereview.chromium.org/1233043003
https://codereview.chromium.org/1230243005
https://codereview.chromium.org/1263053002
https://codereview.chromium.org/1242023005
https://codereview.chromium.org/1280473002
https://codereview.chromium.org/1295163005
https://codereview.chromium.org/1169393003
https://codereview.chromium.org/1298113003

BUG=XWALK-5647