[Android] Fix the error that the screen flashes at startup
The moment when the XWalkViewBridge is added to the XWalkView, the
screen flashes black. The reason is when the SurfaceView appears
in the window the fist time, it requests the window's parameters
changing by calling IWindowSession.relayout(). But if the window
already has appropriate parameters, it will not refresh all the
window's stuff and the screen will not blink. So we add a 0px
SurfaceView at first. This will recreate the window before the
activity is shown on the screen, and when the actual XWalkViewBridge
is added, it will just continue to use the window with current
parameters. The temporary SurfaceView can be removed at last.

BUG=XWALK-4966