[Android] Fixed the bulk data transmission crash on Android
This issue is introduced by patch e5d3b6c2, the solution only works
for Windows but crashes on Android.

On Android as we reuse existing proxy channel when creating extension
servers, the XWalkExtensionServer::OnChannelConnected() will not be
triggered in this case. So we keep a pointer of the channel proxy
instead of a sender pointer when initialize extension server, from
which we can get the renderer process id when we need to retrieve
renderer process and its handler in PostMessageToJSCallback().

BUG=XWALK-5029