[Android] Implement expansion mode
Crosswalk expansion mode allows to load the runtime from an android
expansion file. This is a middle-ground between bundled mode and
shared mode, in that it lets app developers ship the entire app
from their own play store account, but does not count the runtime
size against the 50MB APK size limit.

This is the core functionality for expansion mode, and works for
crosswalk embedding apps that implement expansion file management
themselves. Tools support will be implemented in a follow-up patch.

How to use expansion mode:
1 - Create crosswalk embedding app
2 - Add AndroidManifest metadata named org.xwalk.expansionFile with
    value "main" or "patch" depending on which expansion file to use.
3 - When app starts, crosswalk is loaded from corresponding
    expansion file.
4 - App needs to be aware that expansion file gets ending ".apk"
    appended, so it can be used with DexClassLoader.
5 - App needs to ship assets/icudtl.dat, this will be fixed in the
    future.

BUG=XWALK-4726