[Android] Fix the issue about manifest.json was always used.
This patch is to fix manifest.json was always used issue.
In application's folder, if manifest.json was existed, it will be
always used even if it was not specified in command line.
e.g. if manifest.json was existed and app was packaged by "make_apk.py
--package=org.xwalk.test --name=test --app-root=path-to-app
--app-local-path=index.html", this way won't work, manifest.json will be
used.
Refine the code that user can either specify --manifest or must specify
the --app-* options.

BUG=XWALK-5131