Fix flakiness when building the .NET extensions tests
On a machine where ninja creates a lot of parrallel tasks we were
running in some build errors :
- The .NET extensions tests depends on the .NET bridge but it was
not listed as an explicit dependency.
- In the multiple extensions tests (where we have several extensions in
the same directory with multiple bridges) we were copying multiple times
the bridge and renaming it according to each extension in the directory.
We can't do these tasks in parallel because we copy then rename the bridge for
extension 1 and then in the same time we copy and rename the bridge
for extension 2. Turns out that the bridge file may be accessed by
both actions running in parrallel. We make it sequential : we copy
the bridge for extension 1, then rename it, then copy the bridge for
extension 2 and then rename it.