gyp_xwalk: Sync with gyp_chromium.
We have not synced the gyp_xwalk script with its upstream counterpart
for a few rebase cycles, so there is some catch up to do. This change
imports the following commits to gyp_chromium, making it up-to-date with
Chromium 45.0.2454.93:

6248cdd [iOS] Add a repack steps to iOS upstream for unit tests
013831e Roll src/tools/gyp/ to 29e94a3285ee899d14d5e56a6001682620d3778f.
bf94627 A couple of cleanups for gyp_chromium
9449f96 Disable circular gyp dependency check on Mac
e6f8f83 Put 'tools' directory at front of import path.
0826b63 [WebView] Pack the .pak files for all the locales.