[Win] Presentation API implementation
This patch contains the implementation of the W3C Presentation
API for Crosswalk on Windows.

With the current chromium-crosswalk revision (which is based on M45
upstream version) the Presentation API bindings correspond to
http://www.w3.org/TR/2015/WD-presentation-api-20150701.

Missing API features: there is no bindings for incoming presentaion
sessions and hence message exchange is also not available.

Missing implementation features (TBD in the following patches):
- CSP rules for presentation URL
- UI for presentation window selection