[Android] Fix the issue of cordova 4.x app launch crash.
Provide a separate function for SurfaceView to set visibility.
Keep XWalkView.setVisibility() as an embedding API, the base class's
setVisibility will be called even if xwalk is not ready.
Mark "setSurfaceViewVisibility()" as reserved function to initial later.
In XWalkViewInternal.java, "setVisibility" is just an auxiliary function
to generate XWalkView.setVisibility().

BUG=XWALK-5700