[Android] Upgrade shouldInterceptLoadRequest API
As current shouldInterceptLoadRequest misses lots of important context,
shouldInterceptLoadRequest need to be upgraded to Android L style.
To keep the backward compatibility, the old API still be usable.

So this commit adds the following to shouldInterceptLoadRequest params:
 - isMainFrame
 - hasUserGesture
 - method
 - request headers

And use WebResourceResponse to replace InterceptedRequestData, adds the
following:
 - status code
 - response phrase
 - response headers

The advantage of this implemention compared to android_webview is that it's
independent of the android sdk version, which means all the Crosswalk support
android verisons based application can use the new shouldInterceptLoadRequest
to get the newly added context.

BUG=XWALK-3934, XWALK-4255