[Android] Start generating Maven POM files with the build system.
Currently, the Maven POM XML files part of the Crosswalk Maven
repository in download.01.org are generated by Maven itself when we call
`mvn deploy:deploy-file` with `-DgeneratePom=true`.

While it works for most cases, it is not a flexible solution: if one
needs to include more information in the POM file, such as a dependency
declaration, the POM file cannot be automatically generated.

This patch adds template POM files for both xwalk_core_library and
xwalk_shared_library that are preprocessed during the build and turn
into POM files that can be used with Maven's `install-file` and
`deploy-file` commands.

By default, the artifact IDs used in the POM files are
`xwalk_core_library_canary` and `xwalk_shared_library_canary`, but their
values can be changed with the `xwalk_core_library_artifact_id` and
`xwalk_shared_library_artifact_id` gyp variables.

The new POM files will not be automatically used when building new
Crosswalk releases just yet; the release scripts need to be modified
first, but that will be done shortly.

BUG=XWALK-5092