Roll to Chromium 46.0.2490.86.
This is the latest stable release at the time of writing. This release
was made after Blink was merged into the Chromium repository, so we
levarage the roll to 45.0.2454.101 in master that adjusts the code in
Crosswalk here.

Android and Linux tests seem to be working fine in release mode.