DEPS.xwalk: Re-roll chromium-crosswalk.
No new commits, but a few backports had to have their commit messages
edited. The original crosswalk-16/45.0.2454.101 branch in
chromium-crosswalk has been preserved as
crosswalk-16/45.0.2454.101-with-wrong-cr-commit-position-markers.

After the Blink repository was merged into the Chromium one upstream,
src/build/util/lastchange.py looks for the latest upstream Blink commit
that we use in the devtools URL by going through the commit messages and
finding the first with the "Cr-Commit-Position" marker.

This heuristic works most of the time (100% of the time for upstream),
but it breaks for us if we have backported commits with that marker in
the commit message, as the script will find one of those commits that do
not actually exist upstream. Consequently, remote debugging breaks
because the devtools URL we use does not exist.

Crosswalk 17 is fine, because none of our commits on top of Chromium
contain the "Cr-Commit-Position" marker.