[Windows] Move JavaScript dialog support to app_modal component.
Remove previous code from b19a154 and use the code landed for Linux.

Please note that this commit also fix the ASSERT triggered in debug
build (which is also triggered on Linux). The fix is to make sure
that we initialize the vm state as we're adding transient windows
alongside with the root window.