[Linux][Windows] Implement selection list for autocompleted forms, Part 2
This is part 2 of the implementation for desktop. The patch implements
the actual popup to display the items from the autocompletion.

It works the following way :
- XWalkAutofillPopupController which is a controller for the actual view
implemented in XWalkAutofillPopupView. The view is implemented using
toolkit views.
- XWalkPopupController which contains the logic of the popup itself.

Most of the code is borrowed from upstream however it is very simplified
to meet Crosswalk needs.

BUG=XWALK-4991