[Windows][Linux] Fix geolocation support and implement a system to ask for permissions.
XWalk will now ask for permission using a dialog and the answer will
be kept persistent across runs. XWalkPermissionManager is the entity
responsible to create the dialog asking for permissions and it is also
the one responsible to verify if a permission was granted/denied in the
past.

XWalkContentSettings is the class responsible to store whether a
permission was granted or not and store it onto the filesystem.

Currently there is no way to reset the permissions once granted/denied.
We should probably implement a way to do so (like a command line to
reset the permissions).

Concerning the geolocation, right now the only backend we have is using
the Google Maps API and thus one need to follow these instructions to get
a key https://www.chromium.org/developers/how-tos/api-keys. At the moment
you need GOOGLE_API_KEY, GOOGLE_DEFAULT_CLIENT_ID and
GOOGLE_DEFAULT_CLIENT_SECRET to be defined as environment variables. This
isn't going to work for crosswalk-app-tool : we either need these to be
defined in say the manifest.json or it has to be a command line argument
which could be passed by crosswalk-app-tools.

BUG=XWALK-4989, XWALK-5475